+++ src/triggers/template.ts
@@ -1,6 +1,6 @@
 import { devMessage } from 'src/utils'
 
-const createMeasuredTemplateHook = Hooks.on('createMeasuredTemplate', (template: MeasuredTemplateDocumentPF2e, _options, _id: ChatMessagePF2e['id']) => {
+const createMeasuredTemplateHook = Hooks.on('createMeasuredTemplate', (template: MeasuredTemplateDocumentPF2e) => {
 	const { actor, item, message, flags: { pf2e: { origin } } } = template
 
 	const deliverable = {
@@ -12,8 +12,9 @@ const createMeasuredTemplateHook = Hooks.on('createMeasuredTemplate', (template:
 		item,
 	}
 
-	devMessage('Template Hook Data', deliverable, _options)
-	window.pf2eGraphics.AnimCore.findAndAnimate(deliverable)
+	devMessage('Template Hook Data', deliverable)
+	// Timed out because of some bizzare circumstance where coordinates are not delivered on time resulting in a 0,0 position.
+	setTimeout(() => window.pf2eGraphics.AnimCore.findAndAnimate(deliverable), 100)
 })
 
 if (import.meta.hot) {