+++ src/assets/soundDb.ts
@@ -6,11 +6,21 @@ const database = {
 			1: `${p}/sounds/ovani-sounds/Magic/Shock/Lightning Bolt Impact B.wav`,
 			2: `${p}/sounds/ovani-sounds/Magic/Shock/Lightning Bolt Impact C.wav`,
 		},
-		cast: {
-			0: `${p}/sounds/ovani-sounds/Magic/Shock/Thunderbolt Strike Cast A.wav`,
-			1: `${p}/sounds/ovani-sounds/Magic/Shock/Thunderbolt Strike Cast B.wav`,
-			2: `${p}/sounds/ovani-sounds/Magic/Shock/Thunderbolt Strike Cast C.wav`,
-		},
+	},
+	bow: {
+		0: `${p}/sounds/gamedev-market/BowWhistleShot1.wav`,
+		1: `${p}/sounds/gamedev-market/BowWhistleShot2.wav`,
+		2: `${p}/sounds/gamedev-market/BowWhistleShot3.wav`,
+		3: `${p}/sounds/gamedev-market/BowWhistleShot4.wav`,
+		4: `${p}/sounds/gamedev-market/BowWhistleShot5.wav`,
+		5: `${p}/sounds/gamedev-market/BowWhistleShot6.wav`,
+		6: `${p}/sounds/gamedev-market/BowWhistleShot7.wav`,
+	},
+	crossbow: {
+		0: `${p}/sounds/gamedev-market/Crossbow Shot A.wav`,
+		1: `${p}/sounds/gamedev-market/Crossbow Shot B.wav`,
+		2: `${p}/sounds/gamedev-market/Crossbow Shot C.wav`,
+		3: `${p}/sounds/gamedev-market/Crossbow Shot D.wav`,
 	},
 }